# [WIP] experiment with a v2 serialization format