--- a/fieldpath/serialize.go
+++ b/fieldpath/serialize.go
@@ -18,10 +18,12 @@ package fieldpath
 
 import (
 	"bytes"
+	"fmt"
 	"io"
 	"unsafe"
 
 	jsoniter "github.com/json-iterator/go"
+	"sigs.k8s.io/structured-merge-diff/value"
 )
 
 func (s *Set) ToJSON() ([]byte, error) {
@@ -45,6 +47,30 @@ func (s *Set) ToJSONStream(w io.Writer) error {
 		return err
 	}
 	stream.WriteObjectEnd()
+	return stream.Flush()
+}
+
+func (s *Set) ToJSON_V2Experimental() ([]byte, error) {
+	buf := bytes.Buffer{}
+	err := s.ToJSONStream_V2Experimental(&buf)
+	if err != nil {
+		return nil, err
+	}
+	return buf.Bytes(), nil
+}
+
+func (s *Set) ToJSONStream_V2Experimental(w io.Writer) error {
+	stream := writePool.BorrowStream(w)
+	defer writePool.ReturnStream(stream)
+
+	var r reusableBuilder
+
+	stream.WriteArrayStart()
+	err := s.emitContents_v2(false, stream, &r)
+	if err != nil {
+		return err
+	}
+	stream.WriteArrayEnd()
 	return stream.Flush()
 }
 
@@ -163,18 +189,119 @@ func (s *Set) emitContents_v1(includeSelf bool, stream *jsoniter.Stream, r *reus
 	return manageMemory(stream)
 }
 
+func (s *Set) emitContents_v2(includeSelf bool, stream *jsoniter.Stream, r *reusableBuilder) error {
+	mi, ci := 0, 0
+	first := true
+	preWrite := func() {
+		if first {
+			first = false
+			return
+		}
+		stream.WriteMore()
+	}
+
+	for mi < len(s.Members.members) && ci < len(s.Children.members) {
+		mpe := s.Members.members[mi]
+		cpe := s.Children.members[ci].pathElement
+
+		if mpe.Less(cpe) {
+			preWrite()
+			if err := serializePathElementToStreamV2(stream, mpe); err != nil {
+				return err
+			}
+			stream.WriteMore()
+			stream.WriteArrayStart()
+			stream.WriteArrayEnd()
+			mi++
+		} else if cpe.Less(mpe) {
+			preWrite()
+			if err := serializePathElementToStreamV2(stream, cpe); err != nil {
+				return err
+			}
+			stream.WriteMore()
+			stream.WriteArrayStart()
+			if err := s.Children.members[ci].set.emitContents_v2(false, stream, r); err != nil {
+				return err
+			}
+			stream.WriteArrayEnd()
+			ci++
+		} else {
+			preWrite()
+			if err := serializePathElementToStreamV2(stream, cpe); err != nil {
+				return err
+			}
+			stream.WriteMore()
+			stream.WriteArrayStart()
+			if err := s.Children.members[ci].set.emitContents_v2(true, stream, r); err != nil {
+				return err
+			}
+			stream.WriteArrayEnd()
+			mi++
+			ci++
+		}
+	}
+
+	for mi < len(s.Members.members) {
+		mpe := s.Members.members[mi]
+
+		preWrite()
+		if err := serializePathElementToStreamV2(stream, mpe); err != nil {
+			return err
+		}
+		stream.WriteMore()
+		stream.WriteArrayStart()
+		stream.WriteArrayEnd()
+		mi++
+	}
+
+	for ci < len(s.Children.members) {
+		cpe := s.Children.members[ci].pathElement
+
+		preWrite()
+		if err := serializePathElementToStreamV2(stream, cpe); err != nil {
+			return err
+		}
+		stream.WriteMore()
+		stream.WriteArrayStart()
+		if err := s.Children.members[ci].set.emitContents_v2(false, stream, r); err != nil {
+			return err
+		}
+		stream.WriteArrayEnd()
+		ci++
+	}
+
+	if includeSelf && !first {
+		preWrite()
+		stream.WriteString(".")
+	}
+	return manageMemory(stream)
+}
+
 // FromJSON clears s and reads a JSON formatted set structure.
 func (s *Set) FromJSON(r io.Reader) error {
 	// The iterator pool is completely useless for memory management, grrr.
 	iter := jsoniter.Parse(jsoniter.ConfigCompatibleWithStandardLibrary, r, 4096)
 
-	found, _ := readIter_v1(iter)
-	if found == nil {
-		*s = Set{}
-	} else {
-		*s = *found
+	next := iter.WhatIsNext()
+	switch next {
+	case jsoniter.ObjectValue:
+		found, _ := readIter_v1(iter)
+		if found == nil {
+			*s = Set{}
+		} else {
+			*s = *found
+		}
+		return iter.Error
+	case jsoniter.ArrayValue:
+		found, _ := readIter_v2(iter)
+		if found == nil {
+			*s = Set{}
+		} else {
+			*s = *found
+		}
+		return iter.Error
 	}
-	return iter.Error
+	return fmt.Errorf("expected object or list, got %v", next)
 }
 
 // returns true if this subtree is also (or only) a member of parent; s is nil
@@ -227,6 +354,120 @@ func readIter_v1(iter *jsoniter.Iterator) (children *Set, isMember bool) {
 				*children.Children.Descend(pe) = *grandchildren
 			}
 		}
+		return true
+	})
+	if children == nil {
+		isMember = true
+	}
+
+	return children, isMember
+}
+
+// returns true if this subtree is also (or only) a member of parent; s is nil
+// if there are no further children.
+func readIter_v2(iter *jsoniter.Iterator) (children *Set, isMember bool) {
+	const (
+		KT int = iota
+		KEY
+		BODY
+	)
+	step := KT
+	var keyType byte
+	var pe PathElement
+	iter.ReadArrayCB(func(iter *jsoniter.Iterator) bool {
+		if step == KT {
+			// first is the key type
+			tmp := iter.ReadStringAsSlice()
+			if len(tmp) != 1 {
+				iter.Error = fmt.Errorf("expected a type of key, got %q", string(tmp))
+				return false
+			}
+			if tmp[0] == '.' {
+				isMember = true
+				return true
+			}
+			keyType = tmp[0]
+			if peField[0] != keyType && peValue[0] != keyType && peIndex[0] != keyType && peKey[0] != keyType {
+				iter.Error = fmt.Errorf("expected a type of key, got %q", string(tmp))
+				return false
+			}
+			step = KEY
+			pe = PathElement{}
+			return true
+		}
+
+		if step == KEY {
+			// second is the key contents
+			switch keyType {
+			case peField[0]:
+				str := iter.ReadString()
+				pe.FieldName = &str
+			case peValue[0]:
+				v, err := value.ReadJSONIter(iter)
+				if err != nil {
+					iter.Error = err
+					return false
+				}
+				pe.Value = &v
+			case peKey[0]:
+				v, err := value.ReadJSONIter(iter)
+				if err != nil {
+					iter.Error = err
+					return false
+				}
+				if v.MapValue == nil {
+					iter.Error = fmt.Errorf("expected key value pairs but got %#v", v)
+					return false
+				}
+				pe.Key = v.MapValue
+			case peIndex[0]:
+				i := iter.ReadInt()
+				pe.Index = &i
+			}
+			step = BODY
+			return true
+		}
+
+		if step != BODY {
+			panic("unexpected step type")
+		}
+
+		next := iter.WhatIsNext()
+		if next != jsoniter.ArrayValue {
+			iter.Error = fmt.Errorf("expected a list, got %v", next)
+			return false
+		}
+
+		grandchildren, childIsMember := readIter_v2(iter)
+		if childIsMember {
+			if children == nil {
+				children = &Set{}
+			}
+			m := &children.Members.members
+			// Since we expect that most of the time these will have been
+			// serialized in the right order, we just verify that and append.
+			appendOK := len(*m) == 0 || (*m)[len(*m)-1].Less(pe)
+			if appendOK {
+				*m = append(*m, pe)
+			} else {
+				children.Members.Insert(pe)
+			}
+		}
+		if grandchildren != nil {
+			if children == nil {
+				children = &Set{}
+			}
+			// Since we expect that most of the time these will have been
+			// serialized in the right order, we just verify that and append.
+			m := &children.Children.members
+			appendOK := len(*m) == 0 || (*m)[len(*m)-1].pathElement.Less(pe)
+			if appendOK {
+				*m = append(*m, setNode{pe, grandchildren})
+			} else {
+				*children.Children.Descend(pe) = *grandchildren
+			}
+		}
+		step = KT
 		return true
 	})
 	if children == nil {