--- a/fieldpath/set_test.go
+++ b/fieldpath/set_test.go
@@ -83,9 +83,11 @@ func BenchmarkFieldSet(b *testing.B) {
 		}
 		operands := make([]*Set, 500)
 		serialized := make([][]byte, len(operands))
+		serializedV2 := make([][]byte, len(operands))
 		for i := range operands {
 			operands[i] = makeSet()
 			serialized[i], _ = operands[i].ToJSON()
+			serializedV2[i], _ = operands[i].ToJSON_V2Experimental()
 		}
 		randOperand := func() *Set { return operands[rand.Intn(len(operands))] }
 
@@ -101,6 +103,18 @@ func BenchmarkFieldSet(b *testing.B) {
 				randOperand().Has(randomPathMaker.makePath(here.minPathLen, here.maxPathLen))
 			}
 		})
+
+		b.Run(fmt.Sprintf("serialize-size-%v", here.size), func(b *testing.B) {
+			total := 0
+			for _, b := range serialized {
+				total += len(b)
+			}
+			b.ReportAllocs()
+			// We're just hacking the allocation reporter to report the size.
+			for i := 0; i < b.N; i++ {
+				_ = make([]byte, total)
+			}
+		})
 		b.Run(fmt.Sprintf("serialize-%v", here.size), func(b *testing.B) {
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
@@ -112,6 +126,31 @@ func BenchmarkFieldSet(b *testing.B) {
 			s := NewSet()
 			for i := 0; i < b.N; i++ {
 				s.FromJSON(bytes.NewReader(serialized[rand.Intn(len(serialized))]))
+			}
+		})
+
+		b.Run(fmt.Sprintf("serializeV2-size-%v", here.size), func(b *testing.B) {
+			total := 0
+			for _, b := range serializedV2 {
+				total += len(b)
+			}
+			b.ReportAllocs()
+			// We're just hacking the allocation reporter to report the size.
+			for i := 0; i < b.N; i++ {
+				_ = make([]byte, total)
+			}
+		})
+		b.Run(fmt.Sprintf("serializeV2-%v", here.size), func(b *testing.B) {
+			b.ReportAllocs()
+			for i := 0; i < b.N; i++ {
+				randOperand().ToJSON_V2Experimental()
+			}
+		})
+		b.Run(fmt.Sprintf("deserializeV2-%v", here.size), func(b *testing.B) {
+			b.ReportAllocs()
+			s := NewSet()
+			for i := 0; i < b.N; i++ {
+				s.FromJSON(bytes.NewReader(serializedV2[rand.Intn(len(serialized))]))
 			}
 		})