```
BenchmarkFieldSet/serialize-size-20-12             20000             85121 ns/op          245760 B/op          1 allocs/op
BenchmarkFieldSet/serialize-20-12                 100000             16411 ns/op            2073 B/op         23 allocs/op
BenchmarkFieldSet/deserialize-20-12                50000             39908 ns/op           10965 B/op        215 allocs/op
BenchmarkFieldSet/serializeV2-size-20-12           20000             77776 ns/op          229376 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-20-12               200000             11355 ns/op            1766 B/op         54 allocs/op
BenchmarkFieldSet/deserializeV2-20-12              50000             26733 ns/op            9819 B/op        139 allocs/op
BenchmarkFieldSet/serialize-size-50-12              5000            238835 ns/op          696320 B/op          1 allocs/op
BenchmarkFieldSet/serialize-50-12                  50000             37537 ns/op            5282 B/op         48 allocs/op
BenchmarkFieldSet/deserialize-50-12                20000             84362 ns/op           23456 B/op        618 allocs/op
BenchmarkFieldSet/serializeV2-size-50-12           10000            219826 ns/op          638976 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-50-12                50000             27731 ns/op            6188 B/op        146 allocs/op
BenchmarkFieldSet/deserializeV2-50-12              30000             58751 ns/op           20203 B/op        400 allocs/op
BenchmarkFieldSet/serialize-size-100-12              500           2853465 ns/op         2310144 B/op          1 allocs/op
BenchmarkFieldSet/serialize-100-12                 10000            114115 ns/op           16990 B/op         93 allocs/op
BenchmarkFieldSet/deserialize-100-12                5000            237687 ns/op           70740 B/op       2163 allocs/op
BenchmarkFieldSet/serializeV2-size-100-12            500           2613701 ns/op         2220032 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-100-12               20000             87290 ns/op           17660 B/op        491 allocs/op
BenchmarkFieldSet/deserializeV2-100-12             10000            156089 ns/op           59673 B/op       1424 allocs/op
BenchmarkFieldSet/serialize-size-500-12              100          15700207 ns/op        11509760 B/op          1 allocs/op
BenchmarkFieldSet/serialize-500-12                  3000            598173 ns/op           77419 B/op        470 allocs/op
BenchmarkFieldSet/deserialize-500-12                2000           1163462 ns/op          328855 B/op      10781 allocs/op
BenchmarkFieldSet/serializeV2-size-500-12           1000           2266963 ns/op        11059200 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-500-12                3000            504621 ns/op          108949 B/op       2460 allocs/op
BenchmarkFieldSet/deserializeV2-500-12              2000            721080 ns/op          273557 B/op       7085 allocs/op
BenchmarkFieldSet/serialize-size-1000-12             100          30542226 ns/op        24870912 B/op          1 allocs/op
BenchmarkFieldSet/serialize-1000-12                 1000           1132630 ns/op          173221 B/op        999 allocs/op
BenchmarkFieldSet/deserialize-1000-12               1000           2043855 ns/op          721161 B/op      23545 allocs/op
BenchmarkFieldSet/serializeV2-size-1000-12           200           5231563 ns/op        24002560 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-1000-12               2000           1180381 ns/op          222182 B/op       5393 allocs/op
BenchmarkFieldSet/deserializeV2-1000-12             1000           1491113 ns/op          601350 B/op      15550 allocs/op

```