@lavalamp 2019-08-20
This is 4% smaller than v1, significantly faster to read, and significantly slower to write (but that can almost certainly be fixed).

Cons: ugly