@lavalamp 2019-08-20
In trying to figure out where these allocs are coming from I've made the v1 serialize do 4x fewer allocs...