@lavalamp 2019-08-20
Effect on existing benchmarks:

```
benchmark                                  old ns/op     new ns/op     delta
BenchmarkFieldSet/serialize-20-12          15352         15551         +1.30%
BenchmarkFieldSet/serialize-50-12          37788         33236         -12.05%
BenchmarkFieldSet/serialize-100-12         112891        96619         -14.41%
BenchmarkFieldSet/serialize-500-12         605046        502300        -16.98%
BenchmarkFieldSet/serialize-1000-12        1291668       997306        -22.79%

benchmark                                  old allocs     new allocs     delta
BenchmarkFieldSet/serialize-20-12          23             4              -82.61%
BenchmarkFieldSet/serialize-50-12          48             5              -89.58%
BenchmarkFieldSet/serialize-100-12         94             5              -94.68%
BenchmarkFieldSet/serialize-500-12         470            34             -92.77%
BenchmarkFieldSet/serialize-1000-12        998            213            -78.66%
```


Final perf numbers:

```
BenchmarkFieldSet/serialize-size-20-12     20000             83303 ns/op          245760 B/op          1 allocs/op
BenchmarkFieldSet/serialize-20-12         100000             15551 ns/op            3178 B/op          4 allocs/op
BenchmarkFieldSet/deserialize-20-12        50000             37749 ns/op           10959 B/op        215 allocs/op
BenchmarkFieldSet/serializeV2-size-20-12                   20000             80997 ns/op          229376 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-20-12                       200000              9205 ns/op            3577 B/op          5 allocs/op
BenchmarkFieldSet/deserializeV2-20-12                      50000             24750 ns/op            9818 B/op        139 allocs/op
BenchmarkFieldSet/serialize-size-50-12                      5000            222826 ns/op          696320 B/op          1 allocs/op
BenchmarkFieldSet/serialize-50-12                          50000             33236 ns/op            6990 B/op          5 allocs/op
BenchmarkFieldSet/deserialize-50-12                        20000             80038 ns/op           23436 B/op        617 allocs/op
BenchmarkFieldSet/serializeV2-size-50-12                   10000            206999 ns/op          638976 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-50-12                       100000             22421 ns/op           10298 B/op          7 allocs/op
BenchmarkFieldSet/deserializeV2-50-12                      30000             55440 ns/op           20205 B/op        401 allocs/op
BenchmarkFieldSet/serialize-size-100-12                      500           2778816 ns/op         2301952 B/op          1 allocs/op
BenchmarkFieldSet/serialize-100-12                         20000             96619 ns/op           17437 B/op          5 allocs/op
BenchmarkFieldSet/deserialize-100-12                       10000            236002 ns/op           70630 B/op       2160 allocs/op
BenchmarkFieldSet/serializeV2-size-100-12                   5000            409521 ns/op         2211840 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-100-12                       20000             66361 ns/op           34172 B/op         12 allocs/op
BenchmarkFieldSet/deserializeV2-100-12                     10000            154700 ns/op           59526 B/op       1421 allocs/op
BenchmarkFieldSet/serialize-size-500-12                      100          15555218 ns/op        11517952 B/op          1 allocs/op
BenchmarkFieldSet/serialize-500-12                          3000            502300 ns/op           57654 B/op         34 allocs/op
BenchmarkFieldSet/deserialize-500-12                        2000           1167272 ns/op          329298 B/op      10795 allocs/op
BenchmarkFieldSet/serializeV2-size-500-12                    100          15514742 ns/op        11067392 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-500-12                        5000            379041 ns/op          189197 B/op         32 allocs/op
BenchmarkFieldSet/deserializeV2-500-12                      2000            755009 ns/op          273737 B/op       7093 allocs/op
BenchmarkFieldSet/serialize-size-1000-12                     100          27864919 ns/op        24846336 B/op          1 allocs/op
BenchmarkFieldSet/serialize-1000-12                         2000            997306 ns/op          128769 B/op        213 allocs/op
BenchmarkFieldSet/deserialize-1000-12                       1000           2079805 ns/op          720717 B/op      23530 allocs/op
BenchmarkFieldSet/serializeV2-size-1000-12                   100          13422513 ns/op        23969792 B/op          1 allocs/op
BenchmarkFieldSet/serializeV2-1000-12                       2000           1002787 ns/op          420510 B/op         58 allocs/op
BenchmarkFieldSet/deserializeV2-1000-12                     1000           1769985 ns/op          600941 B/op      15540 allocs/op
```

A possible improvement is to replace the two character code with a single number.