@lavalamp 2019-08-20
TL;DR:
* deserialize 15% faster, 34% fewer allocs
* serialize about even (accidentally sped up existing code, so 20% speed / 95% alloc improvement)
* output 4% smaller, could be improved slightly still
* probably the ugliest output format I've ever made