@lavalamp 2019-08-20
I did my optimization-- it was much bigger than expected, output size is now 25% smaller.

My test uses a bunch of short field names which probably exaggerates the gain.